moved many things no more warnings need to test if actual works, other than that success for update the project

+++ Libs/ServerCore/Models/User/UserPlaytime.cs
@@ -3,5 +3,5 @@
 public class UserPlaytime : UserBase
 {
     public uint UplayId { get; set; }
-    public ulong PlayTime { get; set; }
+    public uint PlayTime { get; set; }
 }
